done the bubble sort

+++ src/sorting/bubble_sort.cpp
@@ -8,7 +8,7 @@ void Bsort(int* arr, int n){
     }
 }
 
-void Bsort_Count(int* arr, int n, int& count){
+void Bsort_Count(int* arr, int n, long long& count){
     for(int i = 0; ++count && i < n - 1; i++){
         for(int j = 0; ++count && j < n - i - 1; j++){
             if(++count && arr[j] > arr[j + 1]) swap(arr[j], arr[j + 1]);
@@ -26,7 +26,7 @@ void Bubble_Sort(int* arr, int n){
 /*
 This function is used to sort the array using bubble sort and count the number of comparisons
 */
-void Bubble_Sort_Count(int* arr, int n, int& count){
+void Bubble_Sort_Count(int* arr, int n, long long& count){
     count = 0;
     Bsort_Count(arr, n, count);
 }
